move download functionality to modal export button
